+++ lib/model/filters/placeholder.dart
@@ -96,7 +96,7 @@ class PlaceholderFilter extends CollectionFilter {
   }
 
   @override
-  Widget? iconBuilder(BuildContext context, double size, {bool showGenericIcon = true}) => Icon(_icon, size: size);
+  Widget? iconBuilder(BuildContext context, double size, {bool allowGenericIcon = true}) => Icon(_icon, size: size);
 
   @override
   String get category => type;