+++ lib/widgets/settings/navigation/drawer.dart
@@ -4,6 +4,7 @@ import 'package:aves/model/settings/settings.dart';
 import 'package:aves/widgets/common/basic/scaffold.dart';
 import 'package:aves/widgets/common/extensions/build_context.dart';
 import 'package:aves/widgets/common/search/page.dart';
+import 'package:aves/widgets/explorer/explorer_page.dart';
 import 'package:aves/widgets/filter_grids/albums_page.dart';
 import 'package:aves/widgets/filter_grids/countries_page.dart';
 import 'package:aves/widgets/filter_grids/places_page.dart';
@@ -41,6 +42,7 @@ class _NavigationDrawerEditorPageState extends State<NavigationDrawerEditorPage>
     CountryListPage.routeName,
     PlaceListPage.routeName,
     TagListPage.routeName,
+    ExplorerPage.routeName,
     SearchPage.routeName,
   };